restore about page folder structure
